Extract Navigator interactor
